Exception handling (#207)
So this PR is a fix for most things that I found which could be fixed without major refactoring, including:
1. Specifying the backup path to a file (via prompt and `--new_path`).
2. KeyboardInterrupts at menus and prompts.
3. Failed but installed commands ran while backing up packages outputted error messages to STDOUT. This was the case for cargo's backup which ran `ls xxx/.cargo/bin` when I didn't have cargo.
4. There was a `configs_mapping` instead of `config_mapping` in `reinstall_configs()` when accessing the config dict.
5. Minor refactoring so that not installed packages won't get a backup directory so we don't need a for loop to clean those up.
6. Reinstalling packages without package backups errored out so now you get prompted to backup packages first.

I tried to follow your lead of allowing users to fix the issues themselves by reprompting and things like that but if you have any further suggestions I'll try my best to do them.

For some of the unfixed problems (critical and otherwise) like failing to reinstall configs for an already existing Sublime Text 3 because copytree() doesn't overwrite, I'll make a separate issue. The main reason for not fixing them in this PR was just because I didn't want to do major code revisions, I just wanted to focus on general handling in this one.

+++ shallow_backup/backup.py
@@ -120,17 +120,16 @@ def backup_packages(backup_path, skip=False):
 	print_pkg_mgr_backup("npm")
 	command = "npm ls --global --parseable=true --depth=0"
 	temp_file_path = "{}/npm_temp_list.txt".format(backup_path)
-	run_cmd_write_stdout(command, temp_file_path)
-	npm_dest_file = "{0}/npm_list.txt".format(backup_path)
-	# Parse npm output
-	with open(temp_file_path, mode="r+") as temp_file:
-		# Skip first line of file
-		temp_file.seek(1)
-		with open(npm_dest_file, mode="w+") as dest:
-			for line in temp_file:
-				dest.write(line.split("/")[-1])
-
-	os.remove(temp_file_path)
+	if not run_cmd_write_stdout(command, temp_file_path):
+		npm_dest_file = "{0}/npm_list.txt".format(backup_path)
+		# Parse npm output
+		with open(temp_file_path, mode="r+") as temp_file:
+			# Skip first line of file
+			temp_file.seek(1)
+			with open(npm_dest_file, mode="w+") as dest:
+				for line in temp_file:
+					dest.write(line.split("/")[-1])
+		os.remove(temp_file_path)
 
 	# atom package manager
 	print_pkg_mgr_backup("Atom")
@@ -150,11 +149,6 @@ def backup_packages(backup_path, skip=False):
 	command = "ls {}".format(applications_path)
 	dest = "{}/system_apps_list.txt".format(backup_path)
 	run_cmd_write_stdout(command, dest)
-
-	# Clean up empty package list files
-	for file in get_abs_path_subfiles(backup_path):
-		if os.path.getsize(file) == 0:
-			os.remove(file)
 
 
 def backup_fonts(backup_path, skip=False):
